chore(dir-structure): move zod.ustils to correct directory
Related work item: GH-19
